refactor(tools): absolute path of assetpath
