+++ packages/tools/src/types.ts
@@ -1,3 +1,4 @@
 export type AssetBoxConfig = {
   assetPaths: string[];
+  filePaths: string[];
 };
